added navbar to homepage ... need to add it to profile page too
